connecting stitcher-splitter to the compilation suite
