+++ prover/example/test_cases/framework_test.go
@@ -14,7 +14,8 @@ import (
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/lookup"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/permutation"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/specialqueries"
-	"github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter"
+	"github.com/consensys/linea-monorepo/prover/protocol/compiler/stitch_split/splitter"
+	"github.com/consensys/linea-monorepo/prover/protocol/compiler/stitch_split/stitcher"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/univariates"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/vortex"
 	"github.com/consensys/linea-monorepo/prover/protocol/ifaces"
@@ -66,7 +67,8 @@ var (
 		innerproduct.Compile,
 	}
 	ARITHMETICS = compilationSuite{
-		splitter.SplitColumns(8),
+		stitcher.Stitcher(4, 8),
+		splitter.Splitter(8),
 		localcs.Compile,
 		globalcs.Compile,
 	}
