+++ prover/example/test_cases/range_test.go
@@ -6,7 +6,8 @@ import (
 	"testing"
 
 	"github.com/consensys/linea-monorepo/prover/maths/common/smartvectors"
-	"github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter"
+	"github.com/consensys/linea-monorepo/prover/protocol/compiler/stitch_split/splitter"
+	"github.com/consensys/linea-monorepo/prover/protocol/compiler/stitch_split/stitcher"
 	"github.com/consensys/linea-monorepo/prover/protocol/wizard"
 )
 
@@ -27,5 +28,6 @@ func proveRange(run *wizard.ProverRuntime) {
 }
 
 func TestRange(t *testing.T) {
-	checkSolved(t, defineRange, proveRange, join(ALL_SPECIALS, compilationSuite{splitter.SplitColumns(8)}, DUMMY), true)
+	checkSolved(t, defineRange, proveRange, join(ALL_SPECIALS,
+		compilationSuite{stitcher.Stitcher(4, 8), splitter.Splitter(8)}, DUMMY), true)
 }
