+++ prover/protocol/compiler/fullrecursion/full_recursion_test.go
@@ -17,8 +17,8 @@ import (
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/mimc"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/permutation"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/specialqueries"
-	"github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter"
-	"github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter/sticker"
+	"github.com/consensys/linea-monorepo/prover/protocol/compiler/stitch_split/splitter"
+	"github.com/consensys/linea-monorepo/prover/protocol/compiler/stitch_split/stitcher"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/univariates"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/vortex"
 	"github.com/consensys/linea-monorepo/prover/protocol/ifaces"
@@ -72,8 +72,8 @@ func TestLookup(t *testing.T) {
 			specialqueries.CompileFixedPermutations,
 			permutation.CompileGrandProduct,
 			innerproduct.Compile,
-			sticker.Sticker(1<<8, 1<<16),
-			splitter.SplitColumns(1 << 16),
+			stitcher.Stitcher(1<<8, 1<<16),
+			splitter.Splitter(1 << 16),
 			localcs.Compile,
 			globalcs.Compile,
 			univariates.CompileLocalOpening,
