+++ prover/protocol/dedicated/functionals/interpolation_test.go
@@ -46,7 +46,8 @@ func TestInterpolate(t *testing.T) {
 		// specialqueries.CompileFixedPermutations,
 		// specialqueries.CompileInclusionPermutations,
 		// innerproduct.Compile,
-		// splitter.SplitColumns(8),
+		// stitcher.Stitcher(4, 8),
+		// splitter.Splitter(8),
 		// arithmetics.CompileLocal,
 		// arithmetics.CompileGlobal,
 		dummy.Compile,
